requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
motor==3.3.2
python-dotenv==1.0.1
cloudinary==1.37.0
pillow==10.2.0
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1 